test(aggregation): Add StrongStationarityProperty (#257)
* Change matrix generation in _inputs.py to be able to generate strongly stationary matrices, strictly weakly stationary matrices and non weakly-stationary matrices
* Change typical matrices to contain the newly added types of matrices
* Rename a few things in _inputs.py to make lines shorter
* Increase tolerance of some existing tests (LinearUnderScalingProperty and test_equivalence_mean of CAGrad) to make them pass on the new typical_matrices
* Add StrongStationarityProperty in _property_testers.py
* Make Constant, DualProj, Mean, Random, Sum and UPGrad extend StrongStationarityProperty.

Note that MGDA should not and does not pass these tests. CAGrad should pass these tests but does not (probably due to some implementation issue of CAGrad). AlignedMTL passes the tests but it's hard to tell if it should in theory. For a few other aggregators, we did not try the property tester, and we do not know if they have the property in theory.
EDIT: this might have changed after changing the dimensions of the matrices that we test.