refactor(autojac): Refactor key requirements (#286)
* Add RequirementError
* Refactor check_and_get_keys into check_keys and update tests accordingly
* Remove the required_keys parameter from Select, Accumulate, _Matrixify and _Reshape and update their instantiations

Note: originally, an idea was to have a method output_keys taking a set of keys as input and returning the output_keys given the provided input_keys (what check_keys does but without the check), and to have a requirement method taking a set of keys as input and returning a boolean as output (whether the given set of input satisfies the requirement of the transform or not). Even though this structure seems better in terms of single responsibility principle, it has the drawbacks of being inefficient, because the requirement method of a Composition would trigger the computation of the requirement method of the child + the computation of the output_keys method of the child. The number of method calls would thus scale quadratically with the number of transforms composed together. With the current structure, this does not happen, which makes it much more efficient with little-to-no downside.
---------

Co-authored-by: Pierre Quinton <pierre.quinton@gmail.com>