Use python 3.13 instead of 3.12
When I suggested Python 3.13.1 instead of 3.12.3 in `CONTRIBUTING.md` (4ed413ee), I should have also changed the python version of build_deploy_documentation.yml to 3.13, but I forgot. The reason is that I think it's better to use the latest Python version whenever we can, whether it be for users, contributors or even github actions. Since this PR will also end up with the removal of build_deploy_documentation.yml, I think it's a good time to also update Python to 3.13 in the new build-deploy-docs.yml file.