test(aggregation): Refactor property testing (#339)
* Transform property testers in _property_testers.py into assertion functions in _asserts.py
* Add n_runs parameter to assert_linear_under_scaling
* Update aggregator tests to parameterize and use the new assertion functions rather than property testers
* Fix dtype issues
* Add test_permutation_invariant to ConFIG and IMTLG
* Update tolerances

* Refactor matrix sampling functions into classes and move them to _matrix_samplers.py
* Add dtype parameter to the constructor of MatrixSamplers
* Add rng parameter to the call method of MatrixSamplers

Note that we decided to have default values for some parameters of the assertion functions (atol, rtol, n_runs and threshold), so that we still have the benefit of having most aggregators tested with the same tolerances, for ease of maintenance (as we don't want to update each aggregator's tolerance individually each time inputs, hardware or cuda drivers change). We have decided that only UPGrad and DualProj would use custom (and tight) tolerances, so that we're aware if we make a regression for them.

For some reason, the permutation_invariance tests now pass on cuda for ConFIG and IMTLG, so they now test it too. This could be due to a change in the cuda implementation of pinv.