refactor: Stop using deprecated typing classes (#353)
* Use Iterable, Sequence, Hashable, Callable from collections.abc rather than typing
* Use contextlib.AbstractContextManager rather than typing.ContextManager

typing.Iterable, typing.Sequence, typing.Callable and typing.ContextManager are deprecated since Python 3.9. typing.Hashable is not deprecated but it's an alias for collections.abc.Hashable, so it's better to use the latter.

Note that typing.TypeAlias is also deprecated, but only since Python 3.12, in favor of using the type statement. We can only make the change when we drop support for Python 3.10 and 3.11.
---------
Co-authored-by: Valérian Rey <valerian.rey@gmail.com>