One-line view_default.
The trick is that remaining_cT_S did not really depend on stride_row, because the floor_division would give the same result if stride_row was not removed. So remaining_cT_S can be precomputed by pre-dividing. Similarly, the modulo can be done in parallel. So there's no need for any for-loop anymore, all can be computed at once.